dapr-distributed-calendar/python/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
dapr==1.12.1
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.1
grpcio==1.60.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
protobuf==4.25.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
typing_extensions==4.9.0
Werkzeug==3.0.1
yarl==1.9.4
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0